# backend-go/docs/swagger/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "auth.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/todoing.api.v1.AuthService/EmailCodeLogin": {
      "post": {
        "summary": "（可选）旧邮箱验证码登录接口 - Deprecated",
        "operationId": "AuthService_EmailCodeLogin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1EmailCodeLoginRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/todoing.api.v1.AuthService/Login": {
      "post": {
        "summary": "用户登录（密码或邮箱验证码）",
        "operationId": "AuthService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/todoing.api.v1.AuthService/Register": {
      "post": {
        "summary": "用户注册",
        "operationId": "AuthService_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RegisterRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/todoing.api.v1.AuthService/SendLoginEmailCode": {
      "post": {
        "summary": "发送登录邮箱验证码",
        "operationId": "AuthService_SendLoginEmailCode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SendLoginEmailCodeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SendLoginEmailCodeRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/todoing.api.v1.AuthService/VerifyToken": {
      "post": {
        "summary": "验证令牌",
        "operationId": "AuthService_VerifyToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1VerifyTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1VerifyTokenRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n\nExample 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\nExample 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\n\nJSON\n\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1EmailCodeLoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      },
      "title": "Deprecated: 独立邮箱验证码登录 (现已合并入 LoginRequest) 保留兼容"
    },
...
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "title": "使用密码登录时必填"
        },
        "emailCodeId": {
          "type": "string",
          "title": "使用邮箱验证码登录: code id"
        },
        "emailCode": {
          "type": "string",
          "title": "使用邮箱验证码登录: code"
        }
      },
      "title": "登录请求（统一: 支持密码或邮箱验证码二选一）"
    },
    "v1LoginResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1Response"
        },
        "token": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/v1User",
          "title": "预留"
        }
      },
      "title": "登录响应"
    },
...
    "v1RegisterRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "title": "必填: 密码注册"
        },
        "captchaId": {
          "type": "string",
          "title": "可选: 若启用图形验证码"
        },
        "captchaCode": {
          "type": "string",
          "title": "可选"
        },
        "emailCodeId": {
          "type": "string",
          "title": "新增: 邮箱验证码注册 (与现有 HTTP /api/auth/send-email-code 对应)"
        },
        "emailCode": {
          "type": "string",
          "title": "新增"
        }
      },
      "title": "注册请求"
    },
    "v1RegisterResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1Response"
        },
        "user": {
          "$ref": "#/definitions/v1User"
        },
        "token": {
          "type": "string",
          "title": "对齐现有返回 (现行 HTTP 直接返回 token)"
        }
      },
      "title": "注册响应"
    },
...
    "v1Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/protobufAny"
        }
      },
      "title": "通用响应结构"
    },
    "v1SendLoginEmailCodeRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "title": "发送登录邮箱验证码请求"
    },
    "v1SendLoginEmailCodeResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1Response"
        }
      }
    },
...
    "v1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "title": "NOTE: 目前后端未单独维护, 用 created_at 占位"
        }
      },
      "title": "用户模型"
    },
    "v1VerifyTokenRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "title": "验证令牌请求"
    },
    "v1VerifyTokenResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1Response"
        },
        "user": {
          "$ref": "#/definitions/v1User"
        }
      },
      "title": "验证令牌响应"
    }
  }
}